Loading Zone Experimentation (Description Important)
Right now toying around with the idea of treating different areas as separate worlds, and the loading zones just being entities on map. Don't know if this will even work but I can get a one way transition so far for it. Breaks if you try to go back to origional world so we will see...